+++ src/_test/core/revert2/mint/revert__mint__0x71.sol
@@ -5,8 +5,8 @@ pragma solidity 0.8.13;
 import "../../../NuggftV1.test.sol";
 
 abstract contract revert__mint__0x71 is NuggftV1Test {
-    uint160 private TOKEN1 = mintable(0);
-    uint160 private TOKEN2 = mintable(1);
+    uint24 private TOKEN1 = mintable(0);
+    uint24 private TOKEN2 = mintable(1);
 
     function test__revert__mint__0x71__fail__desc() public {
         expect.mint().from(users.frank).value(1 ether).exec(TOKEN1);
