+++ src/_test/core/revert2/offer/revert__offer__0x71.sol
@@ -5,12 +5,12 @@ pragma solidity 0.8.13;
 import "../../../NuggftV1.test.sol";
 
 abstract contract revert__offer__0x71 is NuggftV1Test {
-    uint160 private TOKEN1 = mintable(0);
+    uint24 private TOKEN1 = mintable(0);
 
     function test__revert__offer__0x71__fail__desc() public {
         jumpStart();
 
-        uint160 tokenId = nuggft.epoch();
+        uint24 tokenId = nuggft.epoch();
 
         expect.mint().from(users.frank).value(1 ether).exec(TOKEN1);
 
@@ -24,7 +24,7 @@ abstract contract revert__offer__0x71 is NuggftV1Test {
     function test__revert__offer__0x71__pass__desc() public {
         jumpStart();
 
-        uint160 tokenId = nuggft.epoch();
+        uint24 tokenId = nuggft.epoch();
 
         expect.mint().from(users.frank).value(1 ether).exec(TOKEN1);
