+++ src/_test/mock/MockNuggftV1Migrator.sol
@@ -6,7 +6,7 @@ import {INuggftV1Migrator} from "../../interfaces/nuggftv1/INuggftV1Migrator.sol
 
 contract MockNuggftV1Migrator is INuggftV1Migrator {
     function nuggftMigrateFromV1(
-        uint160 tokenId,
+        uint24 tokenId,
         uint256 proof,
         address owner
     ) external payable override {
